move tests to separate folders
- create folders for different types of tests, to enable later running
only certain tests if needed
