+++ package.json
@@ -13,7 +13,7 @@
     "watch": "gulp watch",
     "build": "gulp",
     "prepublish": "gulp",
-    "test": "NODE_ENV=test mocha --reporter spec server/test"
+    "test": "NODE_ENV=test mocha --reporter spec server/test --recursive"
   },
   "engines": {
     "node": ">=4.1.1"
